IndexedDB/tests/submissions/Opera/idbcursor_advance_invalid.htm

<!DOCTYPE html>
<title>IDBCursor.advance() - attempt to call advance twice</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support.js"></script>

<script>

    var db,
      t = async_test();

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        var objStore = db.createObjectStore("test");

        objStore.createIndex("index", "");

        objStore.add("data",  1);
        objStore.add("data2", 2);
        objStore.add("data3", 3);
    };

    open_rq.onsuccess = function(e) {
        var count = 0;
        var cursor_rq = db.transaction("test")
                          .objectStore("test")
                          .index("index")
                          .openCursor();

        cursor_rq.onsuccess = t.step_func(function(e) {
            if (!e.target.result) {
                assert_equals(count, 3, 'count');
                t.done();
                return;
            }
            var cursor = e.target.result;

            cursor.advance(undefined);

            // Second try
            assert_throws('InvalidStateError',
                function() { cursor.advance(); }, 'second advance');

            assert_throws('InvalidStateError',
                function() { cursor.advance(2); }, 'third advance');

            count++;
        });
    };

    test("attempt to pass a key paramenter that is not a valid key", function(e) {
        assert_throws("DataError",
            function() { cursor.advance(document); });

        this.done();
    });

</script>

<div id="log"></div>
